uses mount instead of shallow
